Crate mithril_merkle_tree

Merkle tree and merkelized map primitives used by Mithril nodes.

This crate provides:

MKTree: a Merkle tree backed by a Merkle Mountain Range and a pluggable storer.

MKProof: a proof of membership for a set of leaves in a Merkle tree.

MKMap: a merkelized map whose keys and values are provable and can be nested (recursive proofs).

MKMapProof: a proof of membership for a leaf in a merkelized map (including nested maps).

Modules

test

Test utilities

Structs

MKMap

A map, where the keys and values are merkelized and provable

MKMapProof

A MKMapProof that proves membership of an entry in the merkelized map

MKProof

A Merkle proof

MKTree

A Merkle tree

MKTreeNode

A node of a Merkle tree

MKTreeStoreInMemory

A Merkle tree store in memory

Enums

MKMapNode

A merkelized map node that is used to represent multi layered merkelized map The MKMapNode can be either a MKMap (Merkle map), a MKTree (full Merkle tree) or a MKTreeNode (Merkle tree node, e.g the root of a Merkle tree) Both MKMap and MKTree can generate proofs of membership for elements that they contain, which allows for recursive proof generation for the multiple layers

Traits

MKMapKey

The trait implemented by the keys of a MKMap

MKMapValue

The trait implemented by the values of a MKMap

MKTreeLeafIndexer

The Merkle tree leaves indexer trait

MKTreeStorer

The Merkle tree storer trait

Type Aliases

Bytes

Alias for a byte

MKTreeLeafPosition

Alias for a Merkle tree leaf position

StdError

Generic error type re-exported for convenience.

StdResult

Generic result type re-exported for convenience.