mithril_merkle_tree

Struct MKMap

pub struct MKMap<K: MKMapKey, V: MKMapValue<K>, S: MKTreeStorer> { /* private fields */ }

A map, where the keys and values are merkelized and provable

Implementations

impl<K: MKMapKey, V: MKMapValue<K>, S: MKTreeStorer> MKMap<K, V, S>

pub fn new(entries: &[(K, V)]) -> StdResult<Self>

MKMap factory

pub fn new_from_iter<T: IntoIterator<Item = (K, V)>>( entries: T, ) -> StdResult<Self>

MKMap factory

pub fn keys(&self) -> impl DoubleEndedIterator<Item = &K>

Get the keys of the merkelized map

pub fn insert(&mut self, key: K, value: V) -> StdResult<()>

Insert a new key-value pair Important: keys must be inserted in order to guarantee that the same set of key/values results in the same computation for the root.

pub fn replace(&mut self, key: K, value: V) -> StdResult<()>

Replace the value of an existing key

pub fn contains(&self, leaf: &MKTreeNode) -> Option<(&K, &V)>

Check if the merkelized map contains a leaf (and returns the corresponding key and value if exists)

pub fn get(&self, key: &K) -> Option<&V>

Get the value of the merkelized map for a given key

pub fn iter(&self) -> impl Iterator<Item = (&K, &V)>

Get an iterator for the key and values of the merkelized map

pub fn len(&self) -> usize

Get the length of the merkelized map

pub fn is_empty(&self) -> bool

Check if the merkelized map is empty

pub fn compress(&mut self) -> StdResult<()>

Compress the merkelized map

pub fn compute_root(&self) -> StdResult<MKTreeNode>

Get the root of the merkle tree of the merkelized map

pub fn compute_proof<T: Into<MKTreeNode> + Clone>( &self, leaves: &[T], ) -> StdResult<MKMapProof<K>>

Get the proof for a set of values of the merkelized map (recursively if needed)